# swagger.json
{
   "swagger": "2.0",
   "info": {
     "title": "E-Market",
     "description": "<b>RESPONSES</b>::\n 200-> OK\n   401->Not Authorized\n   201->No Record\n 400->Bad Request \n 500->Internal Server Error",
     
       "version": "1.0"
   },
   "produces": ["application/json"],
   "paths": {
...
     "/api/mobile/profile/getProfile": {
      "get": {
        "x-swagger-router-controller": "/api/mobile/profile/getProfile",
        "operationId": "index",
        "tags": ["Get Profile"],
        "description": "Get Profile",
        "parameters": [
          { "name": "authorization",  "in": "header",  "required":"true" ,"type":"string"}

     

     ],

   
    
    
        "responses": {}
      }
    },
    "/api/mobile/profile/updateProfile": {
      "post": {
        "x-swagger-router-controller": "/api/mobile/profile/updateProfile",
        "operationId": "index",
        "tags": ["Update Profile"],
        "description": "Update Profile",
        "parameters": [
          { "name": "authorization",  "in": "header",  "required":"true" ,"type":"string"},
          { "name": "firstName",  "in": "formData",  "required":true },
          { "name": "lastName",  "in": "formData"},
          { "name": "email",  "in": "formData",  "required":true },
          { "name": "profileImage",  "in": "formData",  "required":true ,"type":"file"},
          { "name": "address",  "in": "formData",  "required":true ,"type":"string"}
     

     ],

   
    
    
        "responses": {}
      }
    },
...
    "/api/mobile/orders/list": {
      "get": {
        "x-swagger-router-controller": "/api/mobile/orders/list",
        "operationId": "index",
        "tags": ["Orders Listing"],
        "description": "Orders Listing",
        "parameters": [
          { "name": "authorization",  "in": "header",  "required":true,"type":"string"},
          { "name": "progressStatus",  "in": "query",  "required":true ,"type":"string"}
        
     ],
        "responses": {}
      }
    },
...
    "/api/mobile/orders/status": {
      "post": {
        "x-swagger-router-controller": "/api/mobile/orders/status",
        "operationId": "index",
        "tags": ["Change Order Status"],
        "description": "Order Status  5->Complete",
        "parameters": [
          { "name": "authorization",  "in": "header",  "required":true,"type":"string"},
          { "name": "id",  "in": "formData",  "required":true ,"type":"string"},
          { "name": "status",  "in": "formData",  "required":true ,"type":"string"}

     ],
        "responses": {}
      }
    },
...
   }
}